Remove pnet_packet module hack
