+++ pnet_macros_support/Cargo.toml
@@ -9,3 +9,6 @@ repository = "https://github.com/libpnet/libpnet"
 description = "Support library for libpnet_macros"
 readme = "../README.md"
 keywords = ["networking", "bitfields", "packet", "protocol"]
+
+[dependencies]
+pnet_base = { path = "../pnet_base", version = ">=0.19.0" }
